Add `Transcript` to `Proof` trait
This commit adds a `Transcript` argument to both the `Proof::prove`
and `Proof::verify` functions to avoid having both `prove` and
`prove_with_transcript` methods for some of the implemented zero
knowledge proofs.

This commit also makes a couple of other minor changes:
* It adds documentation to the `Proof` trait.
* It moves the `rng` argument in `Proof::prove` to the end to be more
consistent with some other uses of `rng` in the library.