Fix security bug in presign protocol
Closes #325.

This commit fixes a security bug in the presign protocol. In the
protocol specification, each round requires that all messages from the
prior round be received (and verified) before proceeding.
Unfortunately, this is not what the code was doing, particular in
round two of the protocol, where it would process a participant's
message as soon as it was received.

This commit fixes that bug, alongside further aligning the presign
implementation with keygen and auxinfo in terms of code structure.

This commit does NOT address any other issues related to a security
audit.